+++ R/helpers.R
@@ -18,13 +18,18 @@ color <- function(message, color) {
   }
 }
 
-isTesting <- function() {
-  identical(Sys.getenv("TEST_JETPACK"), "true")
+enableRenv <- function() {
+  # use load (activate updates profile then calls load)
+  # no need to call quiet since we already set it globally
+  renv::load(renvProject())
 }
 
-enablePackrat <- function() {
-  clean <- !isTesting()
-  suppressMessages(packrat::on(print.banner=FALSE, clean.search.path=clean))
+ensureRepos <- function() {
+  repos <- getOption("repos", list())
+  if (!is.na(repos["CRAN"]) && repos["CRAN"] == "@CRAN@") {
+    repos["CRAN"] <- "https://cloud.r-project.org/"
+    options(repos=repos)
+  }
 }
 
 findDir <- function(path) {
@@ -37,14 +42,18 @@ findDir <- function(path) {
   }
 }
 
-# duplicate logic from
-# packrat:::getDefaultLibPaths()
-getDefaultLibPaths <- function() {
-  strsplit(Sys.getenv("R_PACKRAT_DEFAULT_LIBPATHS", unset=""), .Platform$path.sep, fixed=TRUE)[[1]]
+getDependencies <- function() {
+  renv::dependencies(path=renvProject())
 }
 
 getDesc <- function() {
-  desc::desc(file=packrat::project_dir())
+  desc::desc(file=renvProject())
+}
+
+getMissing <- function(status) {
+  packages <- names(status$lockfile$Package)
+  dependencies <- getDependencies()$Package
+  missing <- setdiff(dependencies, packages)
 }
 
 getName <- function(package) {
@@ -57,7 +66,7 @@ getName <- function(package) {
 
 getStatus <- function(project=NULL) {
   tryCatch({
-    suppressWarnings(packrat::status(project=project, quiet=TRUE))
+    quietly(renv::status(project=project))
   }, error=function(err) {
     msg <- conditionMessage(err)
     if (grepl("This project has not yet been packified", msg)) {
@@ -73,39 +82,10 @@ installHelper <- function(remove=c(), desc=NULL, show_status=FALSE, update_all=F
     desc <- getDesc()
   }
 
-  # configure local repos
-  remotes <- desc$get_remotes()
-  local_repos <- c()
-  bioc <- FALSE
-  for (remote in remotes) {
-    if (startsWith(remote, "local::")) {
-      repo <- dirname(substring(remote, 8))
-      local_repos <- c(local_repos, repo)
-    } else if (startsWith(remote, "bioc::")) {
-      bioc <- TRUE
-    }
-  }
-  packrat::set_opts(local.repos=local_repos, persist=FALSE)
-
-  repos <- getOption("repos")
-  if (bioc && is.na(repos["BioCsoft"])) {
-    # not ideal, will hopefully be fixed with
-    # https://github.com/rstudio/packrat/issues/507
-    bioc_repos <- c(
-      BioCsoft="https://bioconductor.org/packages/3.7/bioc",
-      BioCann="https://bioconductor.org/packages/3.7/data/annotation",
-      BioCexp="https://bioconductor.org/packages/3.7/data/experiment",
-      BioCworkflows="https://bioconductor.org/packages/3.7/workflows"
-    )
-
-    # TODO need to set options(repos=...) as well for the snapshot to work
-    packrat::set_lockfile_metadata(repos=c(repos, bioc_repos))
-  }
-
   # use a temporary directly
   # this way, we don't update DESCRIPTION
   # until we know it was successful
-  dir <- packrat::project_dir()
+  dir <- renvProject()
   temp_desc <- file.path(dir, "DESCRIPTION")
   desc$write(temp_desc)
   # strip trailing whitespace
@@ -114,27 +94,20 @@ installHelper <- function(remove=c(), desc=NULL, show_status=FALSE, update_all=F
 
   # get status
   status <- getStatus(project=dir)
-  missing <- status[is.na(status$library.version), ]
-  restore <- missing[!is.na(missing$packrat.version), ]
-  need <- missing[is.na(missing$packrat.version), ]
-  missing_packrat <- status[is.na(status$packrat.version), ]
-  mismatch <- status[!is.na(status$library.version) & !is.na(status$packrat.version) & status$packrat.version != status$library.version, ]
-
-  # remove mismatch
-  for (name in mismatch$package) {
-    pkgRemove(name)
-  }
+  need <- getMissing(status)
 
   status_updated <- FALSE
 
-  if (nrow(restore) > 0 || nrow(mismatch) > 0) {
-    suppressWarnings(packrat::restore(project=dir, prompt=FALSE))
+  if (!identical(status$library$Packages, status$lockfile$Packages)) {
+    suppressWarnings(renv::restore(project=dir, prompt=FALSE, clean=TRUE))
 
     # non-vendor approach
     # for (i in 1:nrow(restore)) {
     #   row <- restore[i, ]
     #   devtools::install_version(row$package, version=row$version, dependencies=FALSE)
     # }
+
+    status_updated <- TRUE
   }
 
   if (length(remove) > 0) {
@@ -149,46 +122,32 @@ installHelper <- function(remove=c(), desc=NULL, show_status=FALSE, update_all=F
   specificDeps <- deps[startsWith(deps$version, "== "), ]
   if (nrow(specificDeps) > 0) {
     specificDeps$version <- sub("== ", "", specificDeps$version)
-    specificDeps <- merge(specificDeps, status, by="package")
-    mismatch <- specificDeps[is.na(specificDeps$packrat.version) | specificDeps$version != specificDeps$packrat.version, ]
-    if (nrow(mismatch) > 0) {
-      for (i in 1:nrow(mismatch)) {
-        row <- mismatch[i, ]
+    for (i in 1:nrow(specificDeps)) {
+      row <- specificDeps[i, ]
+      currentDep <- status$lockfile$Packages[[row$package]]
+      if (is.null(currentDep) || currentDep$Version != row$version) {
         remotes::install_version(row$package, version=row$version, reload=FALSE)
+        status_updated <- TRUE
       }
     }
-    status_updated <- TRUE
   }
 
   # in case we're missing any deps
   # unfortunately, install_deps doesn't check version requirements
   # https://github.com/r-lib/devtools/issues/1314
-  if (nrow(need) > 0 || length(remove) > 0 || update_all) {
+  if (length(need) > 0 || length(remove) > 0 || update_all) {
     remotes::install_deps(dir, upgrade=update_all, reload=FALSE)
     status_updated <- TRUE
   }
 
-  if (status_updated || any(!status$currently.used)) {
-    suppressMessages(packrat::clean(project=dir))
-    status_updated <- TRUE
-  }
-
-  if (status_updated || length(missing_packrat) > 0) {
-    # Bioconductor packages fail to download source
-    suppressMessages(packrat::snapshot(project=dir, prompt=FALSE, ignore.stale=TRUE, snapshot.sources=FALSE))
-
-    # loaded packages like curl can be missing on Windows
-    # so see if we need to restore again
-    status <- getStatus(project=dir)
-    if (any(is.na(status$library.version))) {
-      suppressWarnings(packrat::restore(project=dir, prompt=FALSE))
-    }
+  if (status_updated) {
+    suppressMessages(renv::snapshot(project=dir, prompt=FALSE))
   }
 
   # copy back after successful
   jetpack_dir <- getOption("jetpack_dir")
-  file.copy(file.path(packrat::project_dir(), "DESCRIPTION"), file.path(jetpack_dir, "DESCRIPTION"), overwrite=TRUE)
-  file.copy(file.path(packrat::project_dir(), "packrat", "packrat.lock"), file.path(jetpack_dir, "packrat.lock"), overwrite=TRUE)
+  file.copy(file.path(renvProject(), "DESCRIPTION"), file.path(jetpack_dir, "DESCRIPTION"), overwrite=TRUE)
+  file.copy(file.path(renvProject(), "renv.lock"), file.path(jetpack_dir, "renv.lock"), overwrite=TRUE)
 
   if (show_status) {
     if (status_updated) {
@@ -199,40 +158,50 @@ installHelper <- function(remove=c(), desc=NULL, show_status=FALSE, update_all=F
   }
 }
 
+isTesting <- function() {
+  identical(Sys.getenv("TEST_JETPACK"), "true")
+}
+
 isWindows <- function() {
   .Platform$OS.type != "unix"
+}
+
+keepwd <- function(code) {
+  wd <- getwd()
+  tryCatch(code, finally={ setwd(wd) })
+}
+
+loadExternal <- function(package) {
+  lib_paths <- getOption("jetpack_lib")
+  loadNamespace(package, lib.loc=lib_paths)
 }
 
 oneLine <- function(x) {
   gsub("\n", " ", x)
 }
 
-noPackrat <- function() {
-  if (packratOn()) {
-    packrat::off()
+noRenv <- function() {
+  if (renvOn()) {
+    renv::deactivate(renvProject())
   }
 }
 
 packified <- function() {
-  file.exists(file.path(packrat::project_dir(), "packrat"))
+  file.exists(file.path(renvProject(), "renv"))
 }
 
 pkgVersion <- function(status, name) {
-  row <- status[status$package == name, ]
-  if (nrow(row) == 0) {
+  row <- status$lockfile$Package[[name]]
+  if (is.null(row)) {
     stop(paste0("Cannot find package '", name, "' in DESCRIPTION file"))
   }
-  row$packrat.version
+  row$Version
 }
 
 pkgRemove <- function(name) {
   if (name %in% rownames(utils::installed.packages())) {
     suppressMessages(utils::remove.packages(name))
   }
-}
-
-packratOn <- function() {
-  !is.na(Sys.getenv("R_PACKRAT_MODE", unset=NA))
 }
 
 prepCommand <- function() {
@@ -247,38 +216,48 @@ prepCommand <- function() {
 
   # copy files
   file.copy(file.path(dir, "DESCRIPTION"), file.path(venv_dir, "DESCRIPTION"), overwrite=TRUE)
-  file.copy(file.path(dir, "packrat.lock"), file.path(venv_dir, "packrat", "packrat.lock"), overwrite=TRUE)
+  file.copy(file.path(dir, "renv.lock"), file.path(venv_dir, "renv.lock"), overwrite=TRUE)
 
-  if (!packratOn()) {
+  if (!renvOn()) {
     if (interactive()) {
-      stop("Packrat must be on to run this. Run:\npackrat::on(); packrat::extlib(\"jetpack\")")
+      stop("renv must be loaded to run this. Restart your R session to continue.")
     } else {
-      enablePackrat()
+      enableRenv()
     }
   }
-
-  packrat::set_opts(use.cache=!isWindows())
 
   ensureRepos()
   checkInsecureRepos()
 }
 
-ensureRepos <- function() {
-  repos <- getOption("repos", list())
-  if (!is.na(repos["CRAN"]) && repos["CRAN"] == "@CRAN@") {
-    repos["CRAN"] <- "https://cloud.r-project.org/"
-    options(repos=repos)
-  }
+quietly <- function(code) {
+  utils::capture.output(suppressMessages({
+    val <- code
+  }))
+  val
+}
+
+renvOn <- function() {
+  !is.na(Sys.getenv("RENV_PROJECT", unset=NA))
+}
+
+renvProject <- function() {
+  getOption("jetpack_venv")
 }
 
 sandbox <- function(code) {
   libs <- c("remotes", "desc", "docopt")
+  for (lib in libs) {
+    loadExternal(lib)
+  }
+  invisible(eval(code))
+}
 
-  if (!interactive()) {
-    suppressMessages(packrat::extlib(libs))
-    invisible(eval(code))
-  } else {
-    invisible(silenceWarnings("so cannot be unloaded", packrat::with_extlib(libs, code)))
+showStatus <- function(status) {
+  packages <- status$library$Packages
+  packages <- packages[order(names(packages))]
+  for (row in packages) {
+    message(paste0("Using ", row$Package, " ", row$Version))
   }
 }
 
@@ -299,11 +278,9 @@ silenceWarnings <- function(msgs, code) {
   res
 }
 
-showStatus <- function(status) {
-  for (i in 1:nrow(status)) {
-    row <- status[i, ]
-    message(paste0("Using ", row$package, " ", row$packrat.version))
-  }
+stopNotMigrated <- function() {
+  cmd <- if (!interactive()) "jetpack migrate" else "jetpack::migrate()"
+  stop(paste0("This project has not yet been migrated to renv.\nRun '", cmd, "' to migrate."))
 }
 
 stopNotPackified <- function() {
@@ -359,7 +336,7 @@ venvDir <- function(dir) {
   }
 
   # TODO better algorithm, but keep dependency free
-  dir_hash <- sum(utf8ToInt(dir))
+  dir_hash <- sum(utf8ToInt(dir)) + 1
   venv_name <- paste0(basename(dir), "-", dir_hash)
   file.path(venv_dir, venv_name)
 }
@@ -368,7 +345,7 @@ setupEnv <- function(dir=getwd(), init=FALSE) {
   ensureRepos()
 
   venv_dir <- venvDir(dir)
-  if (init && file.exists(venv_dir) && !file.exists("packrat.lock")) {
+  if (init && file.exists(venv_dir) && !file.exists(file.path(dir, "renv.lock"))) {
     # remove previous virtual env
     unlink(venv_dir, recursive=TRUE)
   }
@@ -376,20 +353,30 @@ setupEnv <- function(dir=getwd(), init=FALSE) {
     dir.create(venv_dir, recursive=TRUE)
   }
 
-  options(packrat.project.dir=venv_dir)
+  options(renv.verbose=FALSE, renv.config.synchronized.check = FALSE, jetpack_venv=venv_dir, jetpack_lib=.libPaths())
 
-  # initialize packrat
+  # initialize renv
   if (!packified()) {
+    if (file.exists(file.path(dir, "packrat.lock")) && !file.exists(file.path(dir, "renv.lock"))) {
+      stopNotMigrated()
+    }
+
     message("Creating virtual environment...")
 
-    # don't include jetpack in external.packages
-    # since packrat will require it to be installed
-    utils::capture.output(suppressMessages(packrat::init(venv_dir, options=list(print.banner.on.startup=FALSE, use.cache=!isWindows()), enter=FALSE, infer.dependencies=FALSE)))
-    packrat::set_lockfile_metadata(repos=list(CRAN="https://cloud.r-project.org/"))
+    file.copy(file.path(dir, "DESCRIPTION"), file.path(venv_dir, "DESCRIPTION"), overwrite=TRUE)
+
+    # restore wd after init changes it
+    keepwd(quietly(renv::init(project=venv_dir, bare=TRUE, restart=FALSE, settings=list(snapshot.type = "explicit"))))
+    quietly(renv::snapshot(prompt=FALSE, force=TRUE))
+
+    # reload desc
+    if (interactive()) {
+      loadExternal("desc")
+    }
   }
 
-  if (!file.exists("packrat.lock")) {
-    file.copy(file.path(packrat::project_dir(), "packrat", "packrat.lock"), "packrat.lock")
+  if (!file.exists(file.path(dir, "renv.lock"))) {
+    file.copy(file.path(renvProject(), "renv.lock"), file.path(dir, "renv.lock"))
   }
 
   venv_dir