+++ R/install.R
@@ -12,11 +12,11 @@ install <- function(deployment=FALSE) {
 
     if (deployment) {
       status <- getStatus()
-      missing <- status[is.na(status$packrat.version), ]
-      if (nrow(missing) > 0) {
-        stop(paste("Missing packages:", paste(missing$package, collapse=", ")))
+      missing <- getMissing(status)
+      if (length(missing) > 0) {
+        stop(paste("Missing packages:", paste(missing, collapse=", ")))
       }
-      suppressWarnings(packrat::restore(prompt=FALSE))
+      suppressWarnings(renv::restore(prompt=FALSE))
       showStatus(status)
     } else {
       installHelper(show_status=TRUE)