+++ R/load.R
@@ -3,23 +3,34 @@
 #' @export
 #' @keywords internal
 load <- function() {
-  dir <- findDir(getwd())
+  wd <- getwd()
+  dir <- findDir(wd)
 
   if (is.null(dir)) {
     stopNotPackified()
   }
 
-  venv_dir <- setupEnv(dir)
-
-  # must source from virtualenv directory
-  # for RStudio for work properly
-  # this should probably be fixed in Packrat
-  wd <- getwd()
   tryCatch({
-    setwd(venv_dir)
-    utils::capture.output(suppressMessages(source("packrat/init.R")))
-  }, finally={
-    setwd(wd)
+    venv_dir <- setupEnv(dir)
+
+    # must source from virtualenv directory
+    # for RStudio for work properly
+    keepwd({
+      setwd(venv_dir)
+      quietly(source("renv/activate.R"))
+    })
+  }, error = function(e) {
+    msg <- geterrmessage()
+    args <- commandArgs(trailingOnly=TRUE)
+    migrating <- !interactive() && identical(args, "migrate")
+    if (!migrating) {
+      if (interactive()) {
+        stop(e)
+      } else {
+        message(conditionMessage(e))
+        quit()
+      }
+    }
   })
 
   invisible()